[DZ-76] Refactor: 채팅 보내기 로직 함수분리
